Updated ColorTemplate with float parsing

+++ src/kvaddakopter/image_processing/data_types/ColorTemplate.java
@@ -98,17 +98,17 @@ public class ColorTemplate {
 	 * Use this to adapt if no targets are found
 	 */
 	public void adaptToOriginalBounds(){
-		int T = adaptationConstant; // number of updates to 63%
+		float T = adaptationConstant; // number of updates to 63%
 		//Hue update
-		hueLow = (int) (hueLow + ((oHueLow) - hueLow)/T);
-		hueHigh = (int) (hueHigh + ((oHueHigh) - hueHigh)/T);
+		hueLow = (int) ((float)(hueLow + ((oHueLow) - hueLow))/T);
+		hueHigh = (int) ((float)(hueHigh + ((oHueHigh) - hueHigh))/T);
 		
 		//Saturation update
-		saturationLow = (int) (saturationLow +(oSaturationLow - saturationLow)/T);
-		saturationHigh = (int) (saturationHigh + (oSaturationHigh - saturationHigh)/T);
+		saturationLow = (int) ((float)(saturationLow +(oSaturationLow - saturationLow))/T);
+		saturationHigh = (int) ((float)(saturationHigh + (oSaturationHigh - saturationHigh))/T);
 		
 		//Value update
-		valueLow = (int) (valueLow + (oValueLow - valueLow)/T);
-		valueHigh = (int) (valueHigh + (oValueHigh - valueHigh)/T);
+		valueLow = (int) ((float)(valueLow + (oValueLow - valueLow))/T);
+		valueHigh = (int) ((float)(valueHigh + (oValueHigh - valueHigh))/T);
 	}
 }
